change balance in credit to Double - create bad request exception class as parent

+++ src/main/java/com/freshome/dto/credit/CreditCreateDTO.java
@@ -8,6 +8,6 @@ public record CreditCreateDTO(
 
         @NotNull(message = "balance can not be null! ")
         @PositiveOrZero
-        Long balance
+        Double balance
 ) {
 }
